requirements.txt
#__General modules__#
Django==4.2.6
python-decouple==3.8
djangorestframework
Pillow
django-cleanup

#__development modules__#
gunicorn

#__third party modules__#
markdown       
django-filter
drf-yasg[validation]
djangorestframework-simplejwt
django-cors-headers
djoser
django-ckeditor
faker

#__delployment modules__#

#__email third_party modules__#
django-mail-templated
#__refactor and test modules__#
black
flake8
pytest
pytest-django
coreapi

#__random data creator__#
Faker


#__background proccess and cache__#
celery
django-celery-beat
Redis

#__caching modules__#
django-redis

